pyproject.toml
[project]
name = "VAI"
version = "0.1.0"
description = "LLM/Agent development project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "matplotlib-fontja>=1.1.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "albumentations>=1.3",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "scikit-image>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.0",
    "poethepoet>=0.31",
]

# ============================================================
# ruff
# ============================================================
[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"

# ============================================================
# mypy
# ============================================================
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# ============================================================
# pytest
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]  # Root-level package
addopts = "-v --tb=short"

# ============================================================
# poethepoet
# ============================================================
[tool.poe.tasks]
lint = "ruff check . --fix"
format = "ruff format ."
typecheck = "mypy . --exclude tests"
test = "pytest"
all = ["lint", "format", "typecheck", "test"]
